+++ src/Symfony/Bundle/SecurityBundle/Resources/config/security.xml
@@ -52,7 +52,7 @@
                 <argument type="service" id="event_dispatcher" />
             </call>
         </service>
-        <service id="security.authentication.manager.guard" class="Symfony\Component\Security\Core\Authentication\GuardAuthenticationManager">
+        <service id="security.authentication.manager.guard" class="Symfony\Component\Security\Http\Authentication\GuardAuthenticationManager">
             <argument /> <!-- guard authenticators -->
             <argument type="service" id="Symfony\Component\Security\Core\User\UserCheckerInterface" /> <!-- User Checker -->
             <argument>%security.authentication.manager.erase_credentials%</argument>
