+++ src/Symfony/Component/Security/Core/composer.json
@@ -20,7 +20,6 @@
         "symfony/event-dispatcher-contracts": "^1.1|^2",
         "symfony/polyfill-php80": "^1.15",
         "symfony/service-contracts": "^1.1.6|^2",
-        "symfony/security-guard": "^4.4",
         "symfony/deprecation-contracts": "^2.1"
     },
     "require-dev": {
