+++ src/Symfony/Component/Security/Guard/Token/PreAuthenticationGuardToken.php
@@ -11,7 +11,7 @@
 
 namespace Symfony\Component\Security\Guard\Token;
 
-use Symfony\Component\Security\Core\Authentication\Token\AbstractToken;
+use Symfony\Component\Security\Core\Authentication\Token\PreAuthenticationGuardToken as CorePreAuthenticationGuardToken;
 
 /**
  * The token used by the guard auth system before authentication.
@@ -22,52 +22,6 @@ use Symfony\Component\Security\Core\Authentication\Token\AbstractToken;
  *
  * @author Ryan Weaver <ryan@knpuniversity.com>
  */
-class PreAuthenticationGuardToken extends AbstractToken implements GuardTokenInterface
+class PreAuthenticationGuardToken extends CorePreAuthenticationGuardToken implements GuardTokenInterface
 {
-    private $credentials;
-    private $guardProviderKey;
-    private $providerKey;
-
-    /**
-     * @param mixed       $credentials
-     * @param string      $guardProviderKey Unique key that bind this token to a specific AuthenticatorInterface
-     * @param string|null $providerKey      The general provider key (when using with HTTP, this is the firewall name)
-     */
-    public function __construct($credentials, string $guardProviderKey, ?string $providerKey = null)
-    {
-        $this->credentials = $credentials;
-        $this->guardProviderKey = $guardProviderKey;
-        $this->providerKey = $providerKey;
-
-        parent::__construct([]);
-
-        // never authenticated
-        parent::setAuthenticated(false);
-    }
-
-    public function getProviderKey(): ?string
-    {
-        return $this->providerKey;
-    }
-
-    public function getGuardProviderKey()
-    {
-        return $this->guardProviderKey;
-    }
-
-    /**
-     * Returns the user credentials, which might be an array of anything you
-     * wanted to put in there (e.g. username, password, favoriteColor).
-     *
-     * @return mixed The user credentials
-     */
-    public function getCredentials()
-    {
-        return $this->credentials;
-    }
-
-    public function setAuthenticated(bool $authenticated)
-    {
-        throw new \LogicException('The PreAuthenticationGuardToken is *never* authenticated.');
-    }
 }
