+++ src/Symfony/Component/Security/Guard/composer.json
@@ -17,7 +17,7 @@
     ],
     "require": {
         "php": "^7.2.5",
-        "symfony/security-core": "^5.0",
+        "symfony/security-core": "^5.1",
         "symfony/security-http": "^4.4.1|^5.0.1",
         "symfony/polyfill-php80": "^1.15"
     },
