+++ src/Symfony/Component/Security/Http/composer.json
@@ -18,7 +18,7 @@
     "require": {
         "php": "^7.2.5",
         "symfony/deprecation-contracts": "^2.1",
-        "symfony/security-core": "^4.4.8|^5.0.8",
+        "symfony/security-core": "^5.1",
         "symfony/http-foundation": "^4.4.7|^5.0.7",
         "symfony/http-kernel": "^4.4|^5.0",
         "symfony/polyfill-php80": "^1.15",
